use Decoder
This replaces a lot of string slicing. This was tricky to do because Tar uses octal numbers encoded as ascii.
But this should be more efficient and less fragile.

Also the mode is remoddeled. It is really a bunch of bit flags, so the elm representation is not a record of booleans